+++ project/get_next_line_utils.c
@@ -6,7 +6,7 @@
 /*   By: rnomoto <rnomoto@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/05 18:52:59 by rnomoto           #+#    #+#             */
-/*   Updated: 2025/04/06 16:39:22 by rnomoto          ###   ########.fr       */
+/*   Updated: 2025/04/06 16:50:08 by rnomoto          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -40,7 +40,7 @@ void	*ft_calloc(size_t nmemb, size_t size)
 char *double_buffer(char *old_buf, size_t size)
 {
     char *new_buf;
-    int i = 0;
+    size_t i = 0;
 
     new_buf = (char *)ft_calloc(sizeof(char), size * 2);
     if (new_buf == NULL)
